LayoutTests/inspector/runtime/getProperties-expected.txt

Tests for the Runtime.getProperties command.


== Running test suite: Runtime.getProperties
-- Running test case: Runtime.getProperties.Object
Evaluating expression...
Getting own properties...
Properties:
    "foo"        =>  "cat" (string)  [writable | enumerable | configurable | isOwn]
    "__proto__"  =>  "Number" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Array
Evaluating expression...
Getting own properties...
Properties:
    "0"          =>  "red" (string)  [writable | enumerable | configurable | isOwn]
    "1"          =>  "green" (string)  [writable | enumerable | configurable | isOwn]
    "2"          =>  "blue" (string)  [writable | enumerable | configurable | isOwn]
    "__proto__"  =>  "Array" (object array)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Constructor
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "prototype"  =>  "Test" (object)  [isOwn]
    "name"       =>  "Test" (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.BoundConstructor
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "bound Test" (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "class Test { }" (function class)  []
    "boundThis"       =>  null (object null)  []

-- Running test case: Runtime.getProperties.BoundConstructorArguments
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "bound Test" (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "class Test { }" (function class)  []
    "boundThis"       =>  null (object null)  []
    "boundArgs"       =>  "Array" (object array)  []

-- Running test case: Runtime.getProperties.Function
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  3 (number)  [configurable | isOwn]
    "name"       =>  "" (string)  [configurable | isOwn]
    "prototype"  =>  "Object" (object)  [writable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.FunctionNoParameters
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "" (string)  [configurable | isOwn]
    "prototype"  =>  "Object" (object)  [writable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.BoundFunction
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  3 (number)  [configurable | isOwn]
    "name"       =>  "bound " (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "function(a, b, c){}" (function)  []
    "boundThis"       =>  null (object null)  []

-- Running test case: Runtime.getProperties.BoundFunctionWithArguments
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "bound " (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "function(a, b, c){}" (function)  []
    "boundThis"       =>  null (object null)  []
    "boundArgs"       =>  "Array" (object array)  []

-- Running test case: Runtime.getProperties.BoundFunctionNoParameters
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "bound " (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "function(){}" (function)  []
    "boundThis"       =>  null (object null)  []

-- Running test case: Runtime.getProperties.BoundFunctionNoParametersWithArguments
Evaluating expression...
Getting own properties...
Properties:
    "length"     =>  0 (number)  [configurable | isOwn]
    "name"       =>  "bound " (string)  [configurable | isOwn]
    "__proto__"  =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]
Internal Properties:
    "targetFunction"  =>  "function(){}" (function)  []
    "boundThis"       =>  null (object null)  []
    "boundArgs"       =>  "Array" (object array)  []

-- Running test case: Runtime.getProperties.Private.Instance.Parent
Evaluating expression...
Getting own properties...
Properties:
    "#instancePrivateProperty"        =>  "instancePrivatePropertyValue parent" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#parentInstancePrivateProperty"  =>  "parentInstancePrivatePropertyValue" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "instancePublicProperty"          =>  "instancePublicPropertyValue parent" (string)  [writable | enumerable | configurable | isOwn]
    "parentInstancePublicProperty"    =>  "parentInstancePublicPropertyValue" (string)  [writable | enumerable | configurable | isOwn]
    "__proto__"                       =>  "PrivateMembersTestClassParent" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Private.Instance.Child
Evaluating expression...
Getting own properties...
Properties:
    "#instancePrivateProperty"        =>  "instancePrivatePropertyValue parent" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#parentInstancePrivateProperty"  =>  "parentInstancePrivatePropertyValue" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#instancePrivateProperty"        =>  "instancePrivatePropertyValue child" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#childInstancePrivateProperty"   =>  "childInstancePrivatePropertyValue" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "instancePublicProperty"          =>  "instancePublicPropertyValue child" (string)  [writable | enumerable | configurable | isOwn]
    "parentInstancePublicProperty"    =>  "parentInstancePublicPropertyValue" (string)  [writable | enumerable | configurable | isOwn]
    "childInstancePublicProperty"     =>  "childInstancePublicPropertyValue" (string)  [writable | enumerable | configurable | isOwn]
    "__proto__"                       =>  "PrivateMembersTestClassChild" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Private.Constructor.Parent
Evaluating expression...
Getting own properties...
Properties:
    "#classPrivateProperty"          =>  "classPrivatePropertyValue parent" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#parentClassPrivateProperty"    =>  "parentClassPrivatePropertyValue" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "length"                         =>  0 (number)  [configurable | isOwn]
    "prototype"                      =>  "PrivateMembersTestClassParent" (object)  [isOwn]
    "classPublicMethod"              =>  "classPublicMethod() { parent }" (function)  [writable | configurable | isOwn]
    "classPublicGetter"              =>  get "get classPublicGetter() { parent }" (function)  [configurable | isOwn]
    "classPublicGetter"              =>  set undefined (undefined)  [configurable | isOwn]
    "classPublicSetter"              =>  get undefined (undefined)  [configurable | isOwn]
    "classPublicSetter"              =>  set "set classPublicSetter(x) { parent }" (function)  [configurable | isOwn]
    "classPublicGetterSetter"        =>  get "get classPublicGetterSetter() { parent }" (function)  [configurable | isOwn]
    "classPublicGetterSetter"        =>  set "set classPublicGetterSetter(x) { parent }" (function)  [configurable | isOwn]
    "parentClassPublicMethod"        =>  "parentClassPublicMethod() { }" (function)  [writable | configurable | isOwn]
    "parentClassPublicGetter"        =>  get "get parentClassPublicGetter() { }" (function)  [configurable | isOwn]
    "parentClassPublicGetter"        =>  set undefined (undefined)  [configurable | isOwn]
    "parentClassPublicSetter"        =>  get undefined (undefined)  [configurable | isOwn]
    "parentClassPublicSetter"        =>  set "set parentClassPublicSetter(x) { }" (function)  [configurable | isOwn]
    "parentClassPublicGetterSetter"  =>  get "get parentClassPublicGetterSetter() { }" (function)  [configurable | isOwn]
    "parentClassPublicGetterSetter"  =>  set "set parentClassPublicGetterSetter(x) { }" (function)  [configurable | isOwn]
    "toString"                       =>  "toString() { return \"<redacted>\"; }" (function)  [writable | configurable | isOwn]
    "classPublicProperty"            =>  "classPublicPropertyValue parent" (string)  [writable | enumerable | configurable | isOwn]
    "parentClassPublicProperty"      =>  "parentClassPublicPropertyValue" (string)  [writable | enumerable | configurable | isOwn]
    "name"                           =>  "PrivateMembersTestClassParent" (string)  [configurable | isOwn]
    "__proto__"                      =>  "function () {\n    [native code]\n}" (function)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Private.Constructor.Child
Evaluating expression...
Getting own properties...
Properties:
    "#classPrivateProperty"         =>  "classPrivatePropertyValue child" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "#childClassPrivateProperty"    =>  "childClassPrivatePropertyValue" (string)  [writable | enumerable | configurable | isOwn | isPrivate]
    "length"                        =>  0 (number)  [configurable | isOwn]
    "prototype"                     =>  "PrivateMembersTestClassChild" (object)  [isOwn]
    "classPublicMethod"             =>  "classPublicMethod() { child }" (function)  [writable | configurable | isOwn]
    "classPublicGetter"             =>  get "get classPublicGetter() { child }" (function)  [configurable | isOwn]
    "classPublicGetter"             =>  set undefined (undefined)  [configurable | isOwn]
    "classPublicSetter"             =>  get undefined (undefined)  [configurable | isOwn]
    "classPublicSetter"             =>  set "set classPublicSetter(x) { child }" (function)  [configurable | isOwn]
    "classPublicGetterSetter"       =>  get "get classPublicGetterSetter() { child }" (function)  [configurable | isOwn]
    "classPublicGetterSetter"       =>  set "set classPublicGetterSetter(x) { child }" (function)  [configurable | isOwn]
    "childClassPublicMethod"        =>  "childClassPublicMethod() { }" (function)  [writable | configurable | isOwn]
    "childClassPublicGetter"        =>  get "get childClassPublicGetter() { }" (function)  [configurable | isOwn]
    "childClassPublicGetter"        =>  set undefined (undefined)  [configurable | isOwn]
    "childClassPublicSetter"        =>  get undefined (undefined)  [configurable | isOwn]
    "childClassPublicSetter"        =>  set "set childClassPublicSetter(x) { }" (function)  [configurable | isOwn]
    "childClassPublicGetterSetter"  =>  get "get childClassPublicGetterSetter() { }" (function)  [configurable | isOwn]
    "childClassPublicGetterSetter"  =>  set "set childClassPublicGetterSetter(x) { }" (function)  [configurable | isOwn]
    "toString"                      =>  "toString() { return \"<redacted>\"; }" (function)  [writable | configurable | isOwn]
    "classPublicProperty"           =>  "classPublicPropertyValue child" (string)  [writable | enumerable | configurable | isOwn]
    "childClassPublicProperty"      =>  "childClassPublicPropertyValue" (string)  [writable | enumerable | configurable | isOwn]
    "name"                          =>  "PrivateMembersTestClassChild" (string)  [configurable | isOwn]
    "__proto__"                     =>  "<redacted>" (function class)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Private.Prototype.Parent
Evaluating expression...
Getting own properties...
Properties:
    "constructor"                       =>  "<redacted>" (function class)  [writable | configurable | isOwn]
    "instancePublicMethod"              =>  "instancePublicMethod() { parent }" (function)  [writable | configurable | isOwn]
    "instancePublicGetter"              =>  get "get instancePublicGetter() { parent }" (function)  [configurable | isOwn]
    "instancePublicGetter"              =>  set undefined (undefined)  [configurable | isOwn]
    "instancePublicSetter"              =>  get undefined (undefined)  [configurable | isOwn]
    "instancePublicSetter"              =>  set "set instancePublicSetter(x) { parent }" (function)  [configurable | isOwn]
    "instancePublicGetterSetter"        =>  get "get instancePublicGetterSetter() { parent }" (function)  [configurable | isOwn]
    "instancePublicGetterSetter"        =>  set "set instancePublicGetterSetter(x) { parent }" (function)  [configurable | isOwn]
    "parentInstancePublicMethod"        =>  "parentInstancePublicMethod() { }" (function)  [writable | configurable | isOwn]
    "parentInstancePublicGetter"        =>  get "get parentInstancePublicGetter() { }" (function)  [configurable | isOwn]
    "parentInstancePublicGetter"        =>  set undefined (undefined)  [configurable | isOwn]
    "parentInstancePublicSetter"        =>  get undefined (undefined)  [configurable | isOwn]
    "parentInstancePublicSetter"        =>  set "set parentInstancePublicSetter(x) { }" (function)  [configurable | isOwn]
    "parentInstancePublicGetterSetter"  =>  get "get parentInstancePublicGetterSetter() { }" (function)  [configurable | isOwn]
    "parentInstancePublicGetterSetter"  =>  set "set parentInstancePublicGetterSetter(x) { }" (function)  [configurable | isOwn]
    "__proto__"                         =>  "Object" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Private.Prototype.Child
Evaluating expression...
Getting own properties...
Properties:
    "constructor"                      =>  "<redacted>" (function class)  [writable | configurable | isOwn]
    "instancePublicMethod"             =>  "instancePublicMethod() { child }" (function)  [writable | configurable | isOwn]
    "instancePublicGetter"             =>  get "get instancePublicGetter() { child }" (function)  [configurable | isOwn]
    "instancePublicGetter"             =>  set undefined (undefined)  [configurable | isOwn]
    "instancePublicSetter"             =>  get undefined (undefined)  [configurable | isOwn]
    "instancePublicSetter"             =>  set "set instancePublicSetter(x) { child }" (function)  [configurable | isOwn]
    "instancePublicGetterSetter"       =>  get "get instancePublicGetterSetter() { child }" (function)  [configurable | isOwn]
    "instancePublicGetterSetter"       =>  set "set instancePublicGetterSetter(x) { child }" (function)  [configurable | isOwn]
    "childInstancePublicMethod"        =>  "childInstancePublicMethod() { }" (function)  [writable | configurable | isOwn]
    "childInstancePublicGetter"        =>  get "get childInstancePublicGetter() { }" (function)  [configurable | isOwn]
    "childInstancePublicGetter"        =>  set undefined (undefined)  [configurable | isOwn]
    "childInstancePublicSetter"        =>  get undefined (undefined)  [configurable | isOwn]
    "childInstancePublicSetter"        =>  set "set childInstancePublicSetter(x) { }" (function)  [configurable | isOwn]
    "childInstancePublicGetterSetter"  =>  get "get childInstancePublicGetterSetter() { }" (function)  [configurable | isOwn]
    "childInstancePublicGetterSetter"  =>  set "set childInstancePublicGetterSetter(x) { }" (function)  [configurable | isOwn]
    "__proto__"                        =>  "PrivateMembersTestClassParent" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.Promise.Resolved
Evaluating expression...
Getting own properties...
Properties:
    "__proto__"  =>  "Promise" (object)  [writable | configurable | isOwn]
Internal Properties:
    "status"  =>  "fulfilled" (string)  []
    "result"  =>  123 (number)  []

-- Running test case: Runtime.getProperties.Promise.Rejected
Evaluating expression...
Getting own properties...
Properties:
    "__proto__"  =>  "Promise" (object)  [writable | configurable | isOwn]
Internal Properties:
    "status"  =>  "rejected" (string)  []
    "result"  =>  123 (number)  []

-- Running test case: Runtime.getProperties.fetchStart.Object
Evaluating expression...
Getting own properties with fetchStart 5...
Properties:
    "U"  =>  5 (number)  [writable | enumerable | configurable | isOwn]
    "G"  =>  6 (number)  [writable | enumerable | configurable | isOwn]
    "S"  =>  7 (number)  [writable | enumerable | configurable | isOwn]
    "I"  =>  8 (number)  [writable | enumerable | configurable | isOwn]
    "Q"  =>  9 (number)  [writable | enumerable | configurable | isOwn]

-- Running test case: Runtime.getProperties.fetchCount.Object
Evaluating expression...
Getting own properties with fetchCount 5...
Properties:
    "A"          =>  0 (number)  [writable | enumerable | configurable | isOwn]
    "Y"          =>  1 (number)  [writable | enumerable | configurable | isOwn]
    "C"          =>  2 (number)  [writable | enumerable | configurable | isOwn]
    "W"          =>  3 (number)  [writable | enumerable | configurable | isOwn]
    "E"          =>  4 (number)  [writable | enumerable | configurable | isOwn]
    "__proto__"  =>  "Object" (object)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.fetchStartCount.Object
Evaluating expression...
Getting own properties with fetchStart 3 with fetchCount 4...
Properties:
    "W"  =>  3 (number)  [writable | enumerable | configurable | isOwn]
    "E"  =>  4 (number)  [writable | enumerable | configurable | isOwn]
    "U"  =>  5 (number)  [writable | enumerable | configurable | isOwn]
    "G"  =>  6 (number)  [writable | enumerable | configurable | isOwn]

-- Running test case: Runtime.getProperties.fetchStart.Array
Evaluating expression...
Getting own properties with fetchStart 5...
Properties:
    "5"  =>  "U" (string)  [writable | enumerable | configurable | isOwn]
    "6"  =>  "G" (string)  [writable | enumerable | configurable | isOwn]
    "7"  =>  "S" (string)  [writable | enumerable | configurable | isOwn]
    "8"  =>  "I" (string)  [writable | enumerable | configurable | isOwn]
    "9"  =>  "Q" (string)  [writable | enumerable | configurable | isOwn]

-- Running test case: Runtime.getProperties.fetchCount.Array
Evaluating expression...
Getting own properties with fetchCount 5...
Properties:
    "0"          =>  "A" (string)  [writable | enumerable | configurable | isOwn]
    "1"          =>  "Y" (string)  [writable | enumerable | configurable | isOwn]
    "2"          =>  "C" (string)  [writable | enumerable | configurable | isOwn]
    "3"          =>  "W" (string)  [writable | enumerable | configurable | isOwn]
    "4"          =>  "E" (string)  [writable | enumerable | configurable | isOwn]
    "__proto__"  =>  "Array" (object array)  [writable | configurable | isOwn]

-- Running test case: Runtime.getProperties.fetchStartCount.Array
Evaluating expression...
Getting own properties with fetchStart 3 with fetchCount 4...
Properties:
    "3"  =>  "W" (string)  [writable | enumerable | configurable | isOwn]
    "4"  =>  "E" (string)  [writable | enumerable | configurable | isOwn]
    "5"  =>  "U" (string)  [writable | enumerable | configurable | isOwn]
    "6"  =>  "G" (string)  [writable | enumerable | configurable | isOwn]
